Tree: Add function subtree_id and use in cat
